Adding Metric class.
