+++ keras/utils/__init__.py
@@ -5,6 +5,7 @@ from . import data_utils
 from . import io_utils
 from . import conv_utils
 from . import losses_utils
+from . import metrics_utils
 
 # Globally-importable utils.
 from .io_utils import HDF5Matrix
